Move logger to utils folder
